pyproject.toml
[project]
name = "llm-as-a-teacher"
version = "0.1.0"
description = "LLM-based GitHub project evaluator for student projects"
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "gradio>=4.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

